requirements.txt
pillow
pyexiftool
torch
torchvision
torchaudio
ultralytics
rawpy
opencv-python
imageio
pytesseract
transformers
huggingface_hub
nltk